build: working on some recent prs
