--- app/components/header.tsx
@@ -1,196 +1,0 @@
-import { LaptopIcon, MoonIcon, SunIcon } from "@radix-ui/react-icons";
-import { Link } from "@remix-run/react";
-import * as React from "react";
-import { useHydrated } from "remix-utils/use-hydrated";
-
-import {
-	getTheme,
-	setTheme as setSystemTheme,
-} from "@/components/theme-switcher";
-import { Button } from "@/components/ui/button";
-import {
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuItem,
-	DropdownMenuLabel,
-	DropdownMenuSeparator,
-	DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import {
-	Badge,
-	CircleUser,
-	Home,
-	LineChart,
-	Menu,
-	Package,
-	Package2,
-	Search,
-	ShoppingCart,
-	Users,
-} from "lucide-react";
-import {
-	Card,
-	CardContent,
-	CardDescription,
-	CardHeader,
-	CardTitle,
-} from "./ui/card";
-import { Input } from "./ui/input";
-import { Sheet, SheetContent, SheetTrigger } from "./ui/sheet";
-
-export function Header() {
-	const hydrated = useHydrated();
-	const [, rerender] = React.useState({});
-	const setTheme = React.useCallback((theme: string) => {
-		setSystemTheme(theme);
-		rerender({});
-	}, []);
-	const theme = getTheme();
-
-	return (
-		<header className="flex fixed w-full h-14 items-center gap-4 border-b bg-muted/40 px-4 lg:h-[60px] lg:px-6">
-			<Sheet>
-				<SheetTrigger asChild>
-					<Button variant="outline" size="icon" className="shrink-0 md:hidden">
-						<Menu className="h-5 w-5" />
-						<span className="sr-only">Toggle navigation menu</span>
-					</Button>
-				</SheetTrigger>
-				<SheetContent side="left" className="flex flex-col">
-					<nav className="grid gap-2 text-lg font-medium">
-						<Link
-							to="#"
-							className="flex items-center gap-2 text-lg font-semibold"
-						>
-							<Package2 className="h-6 w-6" />
-							<span className="sr-only">Acme Inc</span>
-						</Link>
-						<Link
-							to="#"
-							className="mx-[-0.65rem] flex items-center gap-4 rounded-xl px-3 py-2 text-muted-foreground hover:text-foreground"
-						>
-							<Home className="h-5 w-5" />
-							Dashboard
-						</Link>
-						<Link
-							to="#"
-							className="mx-[-0.65rem] flex items-center gap-4 rounded-xl bg-muted px-3 py-2 text-foreground hover:text-foreground"
-						>
-							<ShoppingCart className="h-5 w-5" />
-							Orders
-							<Badge className="ml-auto flex h-6 w-6 shrink-0 items-center justify-center rounded-full">
-								6
-							</Badge>
-						</Link>
-						<Link
-							to="#"
-							className="mx-[-0.65rem] flex items-center gap-4 rounded-xl px-3 py-2 text-muted-foreground hover:text-foreground"
-						>
-							<Package className="h-5 w-5" />
-							Products
-						</Link>
-						<Link
-							to="#"
-							className="mx-[-0.65rem] flex items-center gap-4 rounded-xl px-3 py-2 text-muted-foreground hover:text-foreground"
-						>
-							<Users className="h-5 w-5" />
-							Customers
-						</Link>
-						<Link
-							to="#"
-							className="mx-[-0.65rem] flex items-center gap-4 rounded-xl px-3 py-2 text-muted-foreground hover:text-foreground"
-						>
-							<LineChart className="h-5 w-5" />
-							Analytics
-						</Link>
-					</nav>
-					<div className="mt-auto">
-						<Card>
-							<CardHeader>
-								<CardTitle>Upgrade to Pro</CardTitle>
-								<CardDescription>
-									Unlock all features and get unlimited access to our support
-									team.
-								</CardDescription>
-							</CardHeader>
-							<CardContent>
-								<Button size="sm" className="w-full">
-									Upgrade
-								</Button>
-							</CardContent>
-						</Card>
-					</div>
-				</SheetContent>
-			</Sheet>
-
-			<DropdownMenu>
-				<DropdownMenuTrigger asChild>
-					<Button variant="secondary" size="icon" className="rounded-full">
-						<CircleUser className="h-5 w-5" />
-						<span className="sr-only">Toggle user menu</span>
-					</Button>
-				</DropdownMenuTrigger>
-				<DropdownMenuContent align="end">
-					<DropdownMenuLabel>My Account</DropdownMenuLabel>
-					<DropdownMenuSeparator />
-					<DropdownMenuItem>Settings</DropdownMenuItem>
-					<DropdownMenuItem>Support</DropdownMenuItem>
-					<DropdownMenuSeparator />
-					<DropdownMenuItem>Logout</DropdownMenuItem>
-				</DropdownMenuContent>
-			</DropdownMenu>
-			<DropdownMenu>
-				<DropdownMenuTrigger asChild>
-					<Button
-						className="w-10 h-10 rounded-full border"
-						size="icon"
-						variant="ghost"
-					>
-						<span className="sr-only">Theme selector</span>
-						{!hydrated ? null : theme === "dark" ? (
-							<MoonIcon />
-						) : theme === "light" ? (
-							<SunIcon />
-						) : (
-							<LaptopIcon />
-						)}
-					</Button>
-				</DropdownMenuTrigger>
-				<DropdownMenuContent className="mt-2">
-					<DropdownMenuLabel>Theme</DropdownMenuLabel>
-					<DropdownMenuSeparator />
-					<DropdownMenuItem asChild>
-						<button
-							type="button"
-							className="w-full"
-							onClick={() => setTheme("light")}
-							aria-selected={theme === "light"}
-						>
-							Light
-						</button>
-					</DropdownMenuItem>
-					<DropdownMenuItem asChild>
-						<button
-							type="button"
-							className="w-full"
-							onClick={() => setTheme("dark")}
-							aria-selected={theme === "dark"}
-						>
-							Dark
-						</button>
-					</DropdownMenuItem>
-					<DropdownMenuItem asChild>
-						<button
-							type="button"
-							className="w-full"
-							onClick={() => setTheme("system")}
-							aria-selected={theme === "system"}
-						>
-							System
-						</button>
-					</DropdownMenuItem>
-				</DropdownMenuContent>
-			</DropdownMenu>
-		</header>
-	);
-}
