+++ app/components/layouts/MainLayout.tsx
@@ -1,278 +1,18 @@
-import {
-	getTheme,
-	setTheme as setSystemTheme,
-} from "@/components/theme-switcher";
-import {
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuItem,
-	DropdownMenuLabel,
-	DropdownMenuSeparator,
-	DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { cn } from "@/lib/styles";
-import { Link } from "@remix-run/react";
-import {
-	Badge,
-	Bell,
-	Check,
-	ChevronsUpDown,
-	Home,
-	LaptopIcon,
-	LineChart,
-	MoonIcon,
-	Package,
-	Package2,
-	Search,
-	ShoppingCart,
-	SunIcon,
-	Users,
-} from "lucide-react";
-import { useCallback, useState } from "react";
-import { useHydrated } from "remix-utils/use-hydrated";
-import { Avatar, AvatarFallback, AvatarImage } from "../ui/avatar";
-import { Button } from "../ui/button";
-import { Card } from "../ui/card";
-import {
-	Command,
-	CommandEmpty,
-	CommandGroup,
-	CommandInput,
-	CommandItem,
-	CommandList,
-} from "../ui/command";
-import { Input } from "../ui/input";
-import { Popover, PopoverContent, PopoverTrigger } from "../ui/popover";
+import Sidebar from "../blocks/sidebar";
+import LayoutHeader from "../blocks/layout-header";
 
 interface MainLayoutProps {
-	title?: string;
-	children: React.ReactNode;
+  children: React.ReactNode;
 }
 
-export default function MainLayout({
-	children,
-	title = "Dashboard",
-}: MainLayoutProps) {
-	const hydrated = useHydrated();
-	const [, rerender] = useState({});
-	const setTheme = useCallback((theme: string) => {
-		setSystemTheme(theme);
-		rerender({});
-	}, []);
-	const theme = getTheme();
-
-	return (
-		<div className="grid min-h-screen w-full p-4 md:grid-cols-[220px_1fr] lg:grid-cols-[320px_1fr]">
-			<div className="hidden bg-muted/40 md:block">
-				<div className="flex h-full m-4 mt-0 border bg-card rounded-md max-h-screen flex-col gap-2">
-					<div className="flex h-14 items-center border-b px-4 lg:h-[60px] lg:px-6">
-						<Link to="/" className="flex items-center gap-2 font-semibold">
-							<Package2 className="h-6 w-6" />
-							<span className="">Acme Inc</span>
-						</Link>
-						<Button variant="outline" size="icon" className="ml-auto h-8 w-8">
-							<Bell className="h-4 w-4" />
-							<span className="sr-only">Toggle notifications</span>
-						</Button>
-					</div>
-					<div className="flex-1">
-						<form className="mx-4 mb-4">
-							<div className="relative">
-								<Search className="absolute left-2.5 top-2.5 h-4 w-4 text-muted-foreground" />
-								<Input
-									type="search"
-									placeholder="Search products..."
-									className="w-full appearance-none bg-background pl-8 shadow-none"
-								/>
-							</div>
-						</form>
-						<nav className="grid items-start px-2 text-sm font-medium lg:px-4">
-							<Link
-								to="/"
-								className="flex items-center gap-3 rounded-lg px-3 py-2 text-muted-foreground transition-all hover:text-primary"
-							>
-								<Home className="h-4 w-4" />
-								Dashboard
-							</Link>
-							<Link
-								to="#"
-								className="flex items-center gap-3 rounded-lg px-3 py-2 text-muted-foreground transition-all hover:text-primary"
-							>
-								<ShoppingCart className="h-4 w-4" />
-								Orders
-								<Badge className="ml-auto flex h-6 w-6 shrink-0 items-center justify-center rounded-full">
-									6
-								</Badge>
-							</Link>
-							<Link
-								to="/inventory"
-								className="flex items-center gap-3 rounded-lg bg-muted px-3 py-2 text-primary transition-all hover:text-primary"
-							>
-								<Package className="h-4 w-4" />
-								Products{" "}
-							</Link>
-							<Link
-								to="#"
-								className="flex items-center gap-3 rounded-lg px-3 py-2 text-muted-foreground transition-all hover:text-primary"
-							>
-								<Users className="h-4 w-4" />
-								Customers
-							</Link>
-							<Link
-								to="#"
-								className="flex items-center gap-3 rounded-lg px-3 py-2 text-muted-foreground transition-all hover:text-primary"
-							>
-								<LineChart className="h-4 w-4" />
-								Analytics
-							</Link>
-						</nav>
-					</div>
-					<div className="mt-auto p-4">
-						<UserComboBox />
-					</div>
-				</div>
-			</div>
-			<div className="flex flex-col gap-y-4">
-				<header className="w-full bg-card border rounded-lg flex items-center justify-between p-4 lg:p-6 mx-auto">
-					<h1 className="text-xl">{title}</h1>
-					<DropdownMenu>
-						<DropdownMenuTrigger asChild>
-							<Button
-								className="w-10 h-10 rounded-full border"
-								size="icon"
-								variant="ghost"
-							>
-								<span className="sr-only">Theme selector</span>
-								{!hydrated ? null : theme === "dark" ? (
-									<MoonIcon />
-								) : theme === "light" ? (
-									<SunIcon />
-								) : (
-									<LaptopIcon />
-								)}
-							</Button>
-						</DropdownMenuTrigger>
-						<DropdownMenuContent className="mt-2">
-							<DropdownMenuLabel>Theme</DropdownMenuLabel>
-							<DropdownMenuSeparator />
-							<DropdownMenuItem asChild>
-								<button
-									type="button"
-									className="w-full"
-									onClick={() => setTheme("light")}
-									aria-selected={theme === "light"}
-								>
-									Light
-								</button>
-							</DropdownMenuItem>
-							<DropdownMenuItem asChild>
-								<button
-									type="button"
-									className="w-full"
-									onClick={() => setTheme("dark")}
-									aria-selected={theme === "dark"}
-								>
-									Dark
-								</button>
-							</DropdownMenuItem>
-							<DropdownMenuItem asChild>
-								<button
-									type="button"
-									className="w-full"
-									onClick={() => setTheme("system")}
-									aria-selected={theme === "system"}
-								>
-									System
-								</button>
-							</DropdownMenuItem>
-						</DropdownMenuContent>
-					</DropdownMenu>
-				</header>
-				<main className="flex flex-1 flex-col">{children}</main>
-			</div>
-		</div>
-	);
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-const frameworks = [
-	{
-		value: "next.js",
-		label: "Next.js",
-	},
-	{
-		value: "sveltekit",
-		label: "SvelteKit",
-	},
-	{
-		value: "nuxt.js",
-		label: "Nuxt.js",
-	},
-	{
-		value: "remix",
-		label: "Remix",
-	},
-	{
-		value: "astro",
-		label: "Astro",
-	},
-];
-
-export function UserComboBox() {
-	const [open, setOpen] = useState(false);
-	const [value, setValue] = useState("");
-
-	return (
-		<Popover open={open} onOpenChange={setOpen}>
-			<PopoverTrigger asChild>
-				<Button
-					variant="outline"
-					role="combobox"
-					aria-expanded={open}
-					className="w-full justify-between py-6"
-				>
-					<div className="flex w-full justify-between items-center">
-						<div className="grid grid-cols-2 grid-rows-2 text-start gap-x-0">
-							<Avatar className="row-span-2">
-								<AvatarImage src="https://github.com/shadcn.png" />
-								<AvatarFallback>CN</AvatarFallback>
-							</Avatar>
-							<h2>User Name</h2>
-							<p className="">User Title</p>
-						</div>
-						<ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-					</div>
-				</Button>
-			</PopoverTrigger>
-			<PopoverContent className="w-full p-0">
-				<Command>
-					<CommandInput placeholder="Search framework..." />
-					<CommandList>
-						<CommandEmpty>No framework found.</CommandEmpty>
-						<CommandGroup>
-							{frameworks.map((framework) => (
-								<CommandItem
-									key={framework.value}
-									value={framework.value}
-									onSelect={(currentValue) => {
-										setValue(currentValue === value ? "" : currentValue);
-										setOpen(false);
-									}}
-								>
-									<Check
-										className={cn(
-											"mr-2 h-4 w-4",
-											value === framework.value ? "opacity-100" : "opacity-0",
-										)}
-									/>
-									{framework.label}
-								</CommandItem>
-							))}
-						</CommandGroup>
-					</CommandList>
-				</Command>
-			</PopoverContent>
-		</Popover>
-	);
+export default function MainLayout({ children }: MainLayoutProps) {
+  return (
+    <div className="grid min-h-screen w-full p-4 md:grid-cols-[220px_1fr] lg:grid-cols-[320px_1fr]">
+      <Sidebar />
+      <div className="flex flex-col gap-y-4">
+        <LayoutHeader />
+        <main className="flex flex-1 flex-col">{children}</main>
+      </div>
+    </div>
+  );
 }
