+++ app/root.tsx
@@ -1,72 +1,71 @@
 import {
-	Links,
-	Meta,
-	Outlet,
-	Scripts,
-	ScrollRestoration,
-	isRouteErrorResponse,
-	useRouteError,
+  Links,
+  Meta,
+  Outlet,
+  Scripts,
+  ScrollRestoration,
+  isRouteErrorResponse,
+  useRouteError,
 } from "@remix-run/react";
 
 import { GlobalPendingIndicator } from "@/components/global-pending-indicator";
-import { Header } from "@/components/header";
 import {
-	ThemeSwitcherSafeHTML,
-	ThemeSwitcherScript,
+  ThemeSwitcherSafeHTML,
+  ThemeSwitcherScript,
 } from "@/components/theme-switcher";
 
 import "./globals.css";
 
 function App({ children }: { children: React.ReactNode }) {
-	return (
-		<ThemeSwitcherSafeHTML lang="en">
-			<head>
-				<meta charSet="utf-8" />
-				<meta name="viewport" content="width=device-width, initial-scale=1" />
-				<Meta />
-				<Links />
-				<ThemeSwitcherScript />
-			</head>
-			<body>
-				<GlobalPendingIndicator />
-				{/* <Header /> */}
-				{children}
-				<ScrollRestoration />
-				<Scripts />
-			</body>
-		</ThemeSwitcherSafeHTML>
-	);
+  return (
+    <ThemeSwitcherSafeHTML lang="en">
+      <head>
+        <meta charSet="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <Meta />
+        <Links />
+        <ThemeSwitcherScript />
+      </head>
+      <body>
+        <GlobalPendingIndicator />
+        {/* <Header /> */}
+        {children}
+        <ScrollRestoration />
+        <Scripts />
+      </body>
+    </ThemeSwitcherSafeHTML>
+  );
 }
 
 export default function Root() {
-	return (
-		<App>
-			<Outlet />
-		</App>
-	);
+  return (
+    <App>
+      <Outlet />
+    </App>
+  );
 }
 
 export function ErrorBoundary() {
-	const error = useRouteError();
-	let status = 500;
-	let message = "An unexpected error occurred.";
-	if (isRouteErrorResponse(error)) {
-		status = error.status;
-		switch (error.status) {
-			case 404:
-				message = "Page Not Found";
-				break;
-		}
-	} else {
-		console.error(error);
-	}
+  const error = useRouteError();
+  let status = 500;
+  let message = "An unexpected error occurred.";
+  if (isRouteErrorResponse(error)) {
+    status = error.status;
+    switch (error.status) {
+      case 404:
+        message = "Page Not Found";
+        break;
+    }
+  } else {
+    console.error(error);
+  }
 
-	return (
-		<App>
-			<div className="container prose py-8">
-				<h1>{status}</h1>
-				<p>{message}</p>
-			</div>
-		</App>
-	);
+  return (
+    <App>
+      <div className="container prose py-8">
+        <h1>{status}</h1>
+        <p>{message}</p>
+      </div>
+    </App>
+  );
 }
