+++ keynotes.md
@@ -1,0 +1,5 @@
+DYNAMIC FILL-UP GRID:
+grid-cols-[320px_1fr]: This part defines the custom configuration for the number of columns in a grid layout. In this case, it sets two columns with specific sizes:
+
+320px: The first column is a fixed width of 320 pixels.
+1fr: The second column takes up the remaining available space (fractional unit), which means it will grow to fill the remaining area of the container.
